+++ src/mouse/linux/mod.rs
@@ -3,7 +3,7 @@ use x11::xlib::*;
 use std::time::Instant;
 use x11::xtest::*;
 use super::{MouseClick, MouseScroll};
-use std::{thread, time::Duration};
+use std::{thread, time::Duration, ptr};
 
 
 pub struct Mouse{